domain model refactor
