+++ src/main/java/com/qunite/api/data/QueueRepository.java
@@ -1,4 +1,4 @@
-package com.qunite.api.repository;
+package com.qunite.api.data;
 
 import com.qunite.api.domain.Queue;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,4 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface QueueRepository extends JpaRepository<Queue, Long> {
     long countByCreatorId(Long id);
+
 }
